+++ openmbvcppinterface/openmbvcppinterface/coilspring.cc
@@ -55,7 +55,7 @@ DOMElement *CoilSpring::writeXMLFile(DOMNode *parent) {
 
 void CoilSpring::createHDF5File() {
   DynamicColoredBody::createHDF5File();
-  data=hdf5Group->createChildObject<H5::VectorSerie<double> >("data")(8);
+  data=hdf5Group->createChildObject<H5::VectorSerie<Float> >("data")(8);
   vector<string> columns;
   columns.emplace_back("Time");
   columns.emplace_back("fromPoint x");
@@ -72,7 +72,7 @@ void CoilSpring::openHDF5File() {
   DynamicColoredBody::openHDF5File();
   if(!hdf5Group) return;
   try {
-    data=hdf5Group->openChildObject<H5::VectorSerie<double> >("data");
+    data=hdf5Group->openChildObject<H5::VectorSerie<Float> >("data");
   }
   catch(...) {
     data=nullptr;
